Add new entrypoint API
